requirements.txt
numpy
opencv-python
Pillow
matplotlib
tqdm
tensorboardX
tensorboard
jittor